feat: 보드 권한유저 추가

+++ src/board/board.service.ts
@@ -4,12 +4,16 @@ import { Repository } from 'typeorm';
 import { CreateBoardDto, UpdateBoardDto } from './dto';
 import { BoardData } from './board.interface';
 import { BoardEntity } from '../entities/board.entity';
+import { BoardAuthorityEntity } from '../entities/board_authority.entity';
 
 @Injectable()
 export class BoardService {
   constructor(
     @InjectRepository(BoardEntity)
     private readonly boardRepository: Repository<BoardEntity>,
+
+    @InjectRepository(BoardAuthorityEntity)
+    private readonly boardAuthorityRepository: Repository<BoardAuthorityEntity>,
   ) {}
 
   //-- 서버 에러 --//
@@ -42,40 +46,50 @@ export class BoardService {
 
   //-- 보드 전체보기 --//
   async getAll(user_id: number): Promise<BoardData[]> {
-    try {
-      return await this.boardRepository.find({
-        where: { user_id: user_id },
-        select: {
-          id: true,
-          name: true,
-          bg_color: true,
-        },
-      });
-    } catch (error) {
-      this.serverError(error);
+    const targetBoards = await this.boardRepository.find({
+      where: { user_id: user_id },
+      select: {
+        id: true,
+        name: true,
+        bg_color: true,
+      },
+    });
+
+    if (!targetBoards.length) {
+      throw new HttpException(
+        { message: '보드를 찾을 수 없습니다.' },
+        HttpStatus.NOT_FOUND,
+      );
     }
+
+    return targetBoards;
   }
 
   //-- 보드 상세보기 --//
   async getBoard(board_id: number): Promise<BoardData> {
-    try {
-      return await this.boardRepository.findOne({
-        where: { id: board_id },
-        select: {
-          id: true,
-          name: true,
-          bg_color: true,
-        },
-      });
-    } catch (error) {
-      this.serverError(error);
+    const targetBoard = await this.boardRepository.findOne({
+      where: { id: board_id },
+      select: {
+        id: true,
+        name: true,
+        bg_color: true,
+      },
+    });
+
+    if (!targetBoard) {
+      throw new HttpException(
+        { message: '보드를 찾을 수 없습니다.' },
+        HttpStatus.NOT_FOUND,
+      );
     }
+    return targetBoard;
   }
 
   //-- 보드 수정하기 --//
   async update(
-    updateBoardDto: UpdateBoardDto,
+    user_id: number,
     board_id: number,
+    updateBoardDto: UpdateBoardDto,
   ): Promise<{ message: string }> {
     const targetBoard = await this.boardRepository.findOne({
       where: { id: board_id },
@@ -87,6 +101,14 @@ export class BoardService {
         HttpStatus.NOT_FOUND,
       );
     }
+
+    if (targetBoard.user_id !== user_id) {
+      throw new HttpException(
+        { message: '수정할 권한이 없습니다.' },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     try {
       Object.assign(targetBoard, updateBoardDto);
       await this.boardRepository.save(targetBoard);
@@ -97,13 +119,53 @@ export class BoardService {
   }
 
   //-- 보드 삭제하기 --//
-  async delete(board_id: number): Promise<{ message: string }> {
+  async delete(
+    user_id: number,
+    board_id: number,
+  ): Promise<{ message: string }> {
     const targetBoard = await this.boardRepository.findOne({
       where: { id: board_id },
     });
+
+    if (targetBoard.user_id !== user_id) {
+      throw new HttpException(
+        { message: '삭제할 권한이 없습니다.' },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     try {
       await this.boardRepository.remove(targetBoard);
       return { message: '보드 삭제에 성공했습니다.' };
+    } catch (error) {
+      this.serverError(error);
+    }
+  }
+
+  //-- 보드 권한유저 추가하기 --//
+  async createBoardAuthority(
+    user_id: number,
+    board_id: number,
+  ): Promise<{ message: string }> {
+    const targetBoard = await this.boardRepository.findOne({
+      where: { id: board_id },
+    });
+
+    if (targetBoard.user_id !== user_id) {
+      throw new HttpException(
+        { message: '권한 유저를 추가할 권한이 없습니다.' },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    const newBoardAuthority = this.boardAuthorityRepository.create({
+      user_id,
+      board_id,
+    });
+
+    try {
+      await this.boardAuthorityRepository.save(newBoardAuthority);
+      return { message: '보드 권한유저 추가에 성공했습니다.' };
     } catch (error) {
       this.serverError(error);
     }
